requirements.txt
playwright
pytest-playwright
pytest-html
playwright-stealth
jsonschema
requests
locust
Flask